refactoring of StreamerOutputModuleBase:  moved serialization code to external base class (which doesn't inherit from edm::one::OutputModule). This allows sharing serialization code    with global and limited output modules.
